app/src/main/java/org/openlmis/core/model/repository/ProgramDataFormRepository.java
/*
 * This program is part of the OpenLMIS logistics management information
 * system platform software.
 *
 * Copyright © 2015 ThoughtWorks, Inc.
 *
 * This program is free software: you can redistribute it and/or modify
 * it under the terms of the GNU Affero General Public License as published
 * by the Free Software Foundation, either version 3 of the License, or
 * (at your option) any later version. This program is distributed in the
 * hope that it will be useful, but WITHOUT ANY WARRANTY; without even the
 * implied warranty of MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.
 * See the GNU Affero General Public License for more details. You should
 * have received a copy of the GNU Affero General Public License along with
 * this program. If not, see http://www.gnu.org/licenses. For additional
 * information contact info@OpenLMIS.org
 */

package org.openlmis.core.model.repository;

import android.content.Context;
import com.google.inject.Inject;
import java.util.List;
import org.openlmis.core.exceptions.LMISException;
import org.openlmis.core.model.ProgramDataForm;
import org.openlmis.core.persistence.GenericDao;

@SuppressWarnings("squid:S1874")
public class ProgramDataFormRepository {

  GenericDao<ProgramDataForm> genericDao;

  @Inject
  public ProgramDataFormRepository(Context context) {
    genericDao = new GenericDao<>(ProgramDataForm.class, context);
  }

  public List<ProgramDataForm> list() throws LMISException {
    return genericDao.queryForAll();
  }

}
